Change session message key from :error to :alert
There's only one message key being used in the routes and views,
to keep the demo very straightforward and simple.

It could be confusing for a newcomer to Syro, or even
Ruby to see something like:

```ruby
session[:error] = "Password updated"
```

This uses the more generic term `alert`, which is used by many other
css and web frameworks to cover the cases of both success and error.